add conditional rendering in App.js to display landing page before player add assets folder in src with logo and landingPage folder
