--- visual-music-p5-react-wrapper/src/components/LandingPage/LandingPage.components.jsx
@@ -1,19 +1,0 @@
-import React, { Component } from 'react';
-import './LandingPage.styles.css';
-
-import {Parallax, ParallaxLayer} from 'react-spring/renderprops-addons';
-
-class LandingPage extends Component {
-  constructor() {
-    super();
-    this.state = {};
-  }
-
-  render() {
-    return (
-      <h2>Working</h2>
-    )
-  }
-}
-
-export default LandingPage;
