+++ visual-music-p5-react-wrapper/src/components/LandingPage/LandingPage.styles.css
@@ -1,0 +1,13 @@
+.landing-page-wrap {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background-image: url('../../assets/LandingPageAssets/bg-image.svg');
+  background-position: center;
+  background-size: cover;
+  background-repeat: no-repeat;
+  background-attachment: fixed;
+  color: #FFFFFF;
+}
